@@ -16,39 +23,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.4",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.3.4",
-    "@angular/compiler": "^4.3.4",
-    "@angular/core": "^4.3.4",
-    "@angular/forms": "^4.3.4",
-    "@angular/http": "^4.3.4",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.3.4",
-    "@angular/platform-browser-dynamic": "^4.3.4",
-    "@angular/router": "^4.3.4",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^2.0.4",
-    "@ngrx/router-store": "^1.2.6",
-    "@ngrx/store": "^2.2.3",
-    "@ngrx/store-devtools": "^3.2.4",
-    "@nguniversal/express-engine": "^1.0.0-beta.2",
-    "angular-calendar": "^0.20.1",
+    "@angular/animations": "^4.4.3",
+    "@angular/cdk": "^2.0.0-beta.11",
+    "@angular/common": "^4.4.3",
+    "@angular/compiler": "^4.4.3",
+    "@angular/core": "^4.4.3",
+    "@angular/flex-layout": "^2.0.0-beta.9",
+    "@angular/forms": "^4.4.3",
+    "@angular/http": "^4.4.3",
+    "@angular/material": "^2.0.0-beta.11",
+    "@angular/platform-browser": "^4.4.3",
+    "@angular/platform-browser-dynamic": "^4.4.3",
+    "@angular/router": "^4.4.3",
+    "@ngrx/effects": "^4.0.5",
+    "@ngrx/entity": "github:ngrx/entity-builds",
+    "@ngrx/router-store": "^4.0.4",
+    "@ngrx/store": "^4.0.3",
+    "@ngrx/store-devtools": "^4.0.0",
+    "@nguniversal/express-engine": "^5.0.0-beta.1",
+    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.1",
+    "angular-calendar": "^0.21.2",
     "core-js": "^2.4.1",
     "crypto-js": "^3.1.9-1",
+    "date-fns": "^1.28.5",
     "express": "^4.15.4",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.4",
-    "reselect": "^3.0.0",
     "rxjs": "^5.4.2",
     "web-animations-js": "^2.2.5",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "^1.3.0",
-    "@angular/compiler-cli": "^4.3.4",
-    "@angular/language-service": "^4.3.4",
-    "@angular/platform-server": "^4.3.4",
+    "@angular/cli": "^1.4.3",
+    "@angular/compiler-cli": "^4.4.3",
+    "@angular/language-service": "^4.4.3",
+    "@angular/platform-server": "^4.4.3",
     "@angularclass/hmr": "^2.1.3",
     "@types/crypto-js": "^3.1.33",
     "@types/express": "^4.0.36",
@@ -66,11 +76,13 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ngrx-store-freeze": "^0.1.9",
+    "ngrx-store-freeze": "^0.2.0",
     "protractor": "~5.1.2",
+    "ts-loader": "^2.3.7",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "~2.4.2",
+    "webpack-node-externals": "^1.6.0"
   },
   "description": "The Task Manager System",
   "main": "index.js",
