@@ -1,42 +1,44 @@
-import {Injectable} from '@angular/core';
-import {Actions, Effect, toPayload} from '@ngrx/effects';
-import {Action} from '@ngrx/store';
-import {Observable} from 'rxjs/Observable';
-import {of} from 'rxjs/observable/of';
-import {TaskService} from '../services';
+import { Injectable } from '@angular/core';
+import { Actions, Effect } from '@ngrx/effects';
+import { Action } from '@ngrx/store';
+import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs/observable/of';
+import { TaskService } from '../services';
+import { Task } from '../domain'
 import * as actions from '../actions/task.action';
+import * as taskHistoryActions from '../actions/task-history.action';
 
 @Injectable()
 export class TaskEffects {
 
   @Effect()
   loadTasksInLists$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.LOAD_IN_LISTS)
-    .map(toPayload)
+    .ofType<actions.LoadTasksInListsAction>(actions.LOAD_IN_LISTS)
+    .map(action => action.payload)
     .mergeMap((taskLists) => {
       return this.service$
         .getByLists(taskLists)
         .map(tasks => new actions.LoadTasksInListsSuccessAction(tasks))
         .catch(err => of(new actions.LoadTasksInListsFailAction(JSON.stringify(err))));
-      }
+    }
     );
 
   @Effect()
   addTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.ADD)
-    .map(toPayload)
+    .ofType<actions.AddTaskAction>(actions.ADD)
+    .map(action => action.payload)
     .switchMap((task) => {
       return this.service$
         .add(task)
         .map(t => new actions.AddTaskSuccessAction(t))
         .catch(err => of(new actions.AddTaskFailAction(JSON.stringify(err))));
-      }
+    }
     );
 
   @Effect()
   updateTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.UPDATE)
-    .map(toPayload)
+    .ofType<actions.UpdateTaskAction>(actions.UPDATE)
+    .map(action => action.payload)
     .switchMap(task => this.service$
       .update(task)
       .map(t => new actions.UpdateTaskSuccessAction(t))
@@ -45,8 +47,8 @@ export class TaskEffects {
 
   @Effect()
   removeTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.DELETE)
-    .map(toPayload)
+    .ofType<actions.DeleteTaskAction>(actions.DELETE)
+    .map(action => action.payload)
     .switchMap(task => this.service$
       .del(task)
       .map(t => new actions.DeleteTaskSuccessAction(t))
@@ -55,8 +57,8 @@ export class TaskEffects {
 
   @Effect()
   completeTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.COMPLETE)
-    .map(toPayload)
+    .ofType<actions.CompleteTaskAction>(actions.COMPLETE)
+    .map(action => action.payload)
     .switchMap(task => this.service$
       .complete(task)
       .map(t => new actions.CompleteTaskSuccessAction(t))
@@ -65,9 +67,9 @@ export class TaskEffects {
 
   @Effect()
   moveTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.MOVE)
-    .map(toPayload)
-    .switchMap(({taskId, taskListId}) => this.service$
+    .ofType<actions.MoveTaskAction>(actions.MOVE)
+    .map(action => action.payload)
+    .switchMap(({ taskId, taskListId }) => this.service$
       .move(taskId, taskListId)
       .map(task => new actions.MoveTaskSuccessAction(task))
       .catch(err => of(new actions.MoveTaskFailAction(JSON.stringify(err))))
@@ -75,9 +77,9 @@ export class TaskEffects {
 
   @Effect()
   moveAllTask$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.MOVE_ALL)
-    .map(toPayload)
-    .switchMap(({srcListId, targetListId}) => this.service$.moveAll(srcListId, targetListId)
+    .ofType<actions.MoveAllAction>(actions.MOVE_ALL)
+    .map(action => action.payload)
+    .switchMap(({ srcListId, targetListId }) => this.service$.moveAll(srcListId, targetListId)
       .map(tasks => new actions.MoveAllSuccessAction(tasks))
       .catch(err => of(new actions.MoveAllFailAction(err)))
     );
@@ -89,5 +91,5 @@ export class TaskEffects {
    * @param store$ 注入 redux store
    */
   constructor(private actions$: Actions,
-              private service$: TaskService) {}
+    private service$: TaskService) { }
 }
