@@ -1,157 +1,151 @@
 import {Action} from '@ngrx/store';
-import {type} from '../utils/type.util';
 import {Project, User} from '../domain';
 
-export const ActionTypes = {
-  ADD: type('[Project] Add'),
-  ADD_SUCCESS: type('[Project] Add Success'),
-  ADD_FAIL: type('[Project] Add Fail'),
-  UPDATE: type('[Project] Update'),
-  UPDATE_SUCCESS: type('[Project] Update Success'),
-  UPDATE_FAIL: type('[Project] Update Fail'),
-  UPDATE_LISTS: type('[Project] Update Lists'),
-  UPDATE_LISTS_SUCCESS: type('[Project] Update Lists Success'),
-  UPDATE_LISTS_FAIL: type('[Project] Update Lists Fail'),
-  DELETE: type('[Project] Delete'),
-  DELETE_SUCCESS: type('[Project] Delete Success'),
-  DELETE_FAIL: type('[Project] Delete Fail'),
-  LOADS: type('[Project] Load'),
-  LOADS_SUCCESS: type('[Project] Load Success'),
-  LOADS_FAIL: type('[Project] Load Fail'),
-  SELECT: type('[Project] Select Project'),
-  INVITE: type('[Project] Invite Members'),
-  INVITE_SUCCESS: type('[Project] Invite Members Success'),
-  INVITE_FAIL: type('[Project] Invite Members Fail'),
-};
+export const ADD = '[Project] Add';
+export const ADD_SUCCESS = '[Project] Add Success';
+export const ADD_FAIL = '[Project] Add Fail';
+export const UPDATE = '[Project] Update';
+export const UPDATE_SUCCESS = '[Project] Update Success';
+export const UPDATE_FAIL = '[Project] Update Fail';
+export const UPDATE_LISTS = '[Project] Update Lists';
+export const UPDATE_LISTS_SUCCESS = '[Project] Update Lists Success';
+export const UPDATE_LISTS_FAIL = '[Project] Update Lists Fail';
+export const DELETE = '[Project] Delete';
+export const DELETE_SUCCESS = '[Project] Delete Success';
+export const DELETE_FAIL = '[Project] Delete Fail';
+export const LOADS = '[Project] Load';
+export const LOADS_SUCCESS = '[Project] Load Success';
+export const LOADS_FAIL = '[Project] Load Fail';
+export const SELECT = '[Project] Select Project';
+export const INVITE = '[Project] Invite Members';
+export const INVITE_SUCCESS = '[Project] Invite Members Success';
+export const INVITE_FAIL = '[Project] Invite Members Fail';
 
 export class AddProjectAction implements Action {
-  type = ActionTypes.ADD;
+  readonly type = ADD;
 
   constructor(public payload: Project) {
   }
 }
 
 export class AddProjectSuccessAction implements Action {
-  type = ActionTypes.ADD_SUCCESS;
+  readonly type = ADD_SUCCESS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class AddProjectFailAction implements Action {
-  type = ActionTypes.ADD_FAIL;
+  readonly type = ADD_FAIL;
 
   constructor(public payload: string) {
   }
 }
 
 export class UpdateProjectAction implements Action {
-  type = ActionTypes.UPDATE;
+  readonly type = UPDATE;
 
   constructor(public payload: Project) {
   }
 }
 
 export class UpdateProjectSuccessAction implements Action {
-  type = ActionTypes.UPDATE_SUCCESS;
+  readonly type = UPDATE_SUCCESS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class UpdateProjectFailAction implements Action {
-  type = ActionTypes.UPDATE_FAIL;
+  readonly type = UPDATE_FAIL;
 
   constructor(public payload: string) {
   }
 }
 
 export class UpdateListsAction implements Action {
-  type = ActionTypes.UPDATE_LISTS;
+  readonly type = UPDATE_LISTS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class UpdateListsSuccessAction implements Action {
-  type = ActionTypes.UPDATE_LISTS_SUCCESS;
+  readonly type = UPDATE_LISTS_SUCCESS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class UpdateListsFailAction implements Action {
-  type = ActionTypes.UPDATE_LISTS_FAIL;
+  readonly type = UPDATE_LISTS_FAIL;
 
   constructor(public payload: string) {
   }
 }
 
 export class DeleteProjectAction implements Action {
-  type = ActionTypes.DELETE;
+  readonly type = DELETE;
 
   constructor(public payload: Project) {
   }
 }
 
 export class DeleteProjectSuccessAction implements Action {
-  type = ActionTypes.DELETE_SUCCESS;
+  readonly type = DELETE_SUCCESS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class DeleteProjectFailAction implements Action {
-  type = ActionTypes.DELETE_FAIL;
+  readonly type = DELETE_FAIL;
 
   constructor(public payload: string) {
   }
 }
 
 export class LoadProjectsAction implements Action {
-  type = ActionTypes.LOADS;
-
-  constructor(public payload: any) {
-  }
+  readonly type = LOADS;
 }
 
 export class LoadProjectsSuccessAction implements Action {
-  type = ActionTypes.LOADS_SUCCESS;
+  readonly type = LOADS_SUCCESS;
 
   constructor(public payload: Project[]) {
   }
 }
 
 export class LoadProjectsFailAction implements Action {
-  type = ActionTypes.LOADS_FAIL;
+  readonly type = LOADS_FAIL;
 
   constructor(public payload: string) {
   }
 }
 
 export class SelectProjectAction implements Action {
-  type = ActionTypes.SELECT;
+  readonly type = SELECT;
 
   constructor(public payload: Project) {
   }
 }
 
 export class InviteMembersAction implements Action {
-  type = ActionTypes.INVITE;
+  readonly type = INVITE;
 
   constructor(public payload: {projectId: string; members: User[]}) {
   }
 }
 
 export class InviteMembersSuccessAction implements Action {
-  type = ActionTypes.INVITE_SUCCESS;
+  readonly type = INVITE_SUCCESS;
 
   constructor(public payload: Project) {
   }
 }
 
 export class InviteMembersFailAction implements Action {
-  type = ActionTypes.INVITE_FAIL;
+  readonly type = INVITE_FAIL;
 
   constructor(public payload: string) {
   }
