@@ -1,12 +1,9 @@
 import {Action} from '@ngrx/store';
-import {type} from '../utils/type.util';
 
-export const ActionTypes = {
-  SWITCH_THEME: type('[THEME] Switch Theme'),
-};
+export const SWITCH_THEME = '[THEME] Switch Theme';
 
 export class SwitchThemeAction implements Action {
-  type = ActionTypes.SWITCH_THEME;
+  readonly type = SWITCH_THEME;
 
   constructor(public payload: boolean) {
   }
