@@ -3,12 +3,9 @@ import {NgModule} from '@angular/core';
 import {CoreModule} from './core';
 import {SharedModule} from './shared';
 import {LoginModule} from './login';
-import {AppComponent} from './app.component';
+import {AppComponent} from './core/containers/app';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-  ],
   imports: [
     BrowserModule.withServerTransition({appId: 'taskmgr'}),
     SharedModule,
