@@ -1,8 +1,7 @@
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
-
 import { AppModule } from './app.module';
-import { AppComponent } from './app.component';
+import { AppComponent } from './core/containers/app';
 
 /**
  * Used for server rendering
@@ -10,7 +9,7 @@ import { AppComponent } from './app.component';
 @NgModule({
   imports: [
     AppModule,
-    ServerModule,
+    ServerModule
   ],
   bootstrap: [
     AppComponent
