@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
 import {AuthGuardService} from '../services';
-import {PageNotFoundComponent} from './page-not-found';
+import {PageNotFoundComponent} from './containers/page-not-found';
 
 const routes: Routes = [
   {
