@@ -1,15 +1,16 @@
 import {Directive, Input, Output, EventEmitter, HostListener, ElementRef, Renderer2} from '@angular/core';
 import {DragDropService, DragData} from '../drag-drop.service';
+import { Observable } from 'rxjs/Observable';
 
 @Directive({
-  selector: '[app-droppable][dropTags][dragEnterClass]',
+  selector: '[appDroppable][dropTags][dragEnterClass]',
 })
 export class DropDirective {
 
   @Output() dropped: EventEmitter<DragData> = new EventEmitter();
   @Input() dropTags: string[] = [];
   @Input() dragEnterClass = '';
-  private drag$;
+  private drag$: Observable<DragData | null>;
 
   constructor(
     private el: ElementRef,
@@ -24,8 +25,10 @@ export class DropDirective {
     ev.stopPropagation();
     if (this.el.nativeElement === ev.target) {
       this.drag$.subscribe(dragData => {
-        if (this.dropTags.indexOf(dragData.tag) > -1) {
+        if (this.dropTags.indexOf(dragData!.tag) > -1) {
           this.rd.addClass(this.el.nativeElement, this.dragEnterClass);
+          this.rd.setProperty(this.el.nativeElement, 'dataTransfer.effectAllowed', 'all');
+          this.rd.setProperty(this.el.nativeElement, 'dataTransfer.dropEffect', 'move');
         }
       });
     }
@@ -37,7 +40,7 @@ export class DropDirective {
     ev.stopPropagation();
     if (this.el.nativeElement === ev.target) {
       this.drag$.subscribe(dragData => {
-        if (this.dropTags.indexOf(dragData.tag) > -1) {
+        if (this.dropTags.indexOf(dragData!.tag) > -1) {
           this.rd.setProperty(ev, 'dataTransfer.effectAllowed', 'all');
           this.rd.setProperty(ev, 'dataTransfer.dropEffect', 'move');
         } else {
@@ -54,7 +57,7 @@ export class DropDirective {
     ev.stopPropagation();
     if (this.el.nativeElement === ev.target) {
       this.drag$.subscribe(dragData => {
-        if (this.dropTags.indexOf(dragData.tag) > -1) {
+        if (this.dropTags.indexOf(<string>dragData!.tag) > -1) {
           this.rd.removeClass(this.el.nativeElement, this.dragEnterClass);
         }
       });
@@ -67,9 +70,9 @@ export class DropDirective {
     ev.stopPropagation();
     if (this.el.nativeElement === ev.target) {
       this.drag$.subscribe(dragData => {
-        if (this.dropTags.indexOf(dragData.tag) > -1) {
+        if (this.dropTags.indexOf(dragData!.tag) > -1) {
           this.rd.removeClass(this.el.nativeElement, this.dragEnterClass);
-          this.dropped.emit(dragData);
+          this.dropped.emit(dragData!);
           this.service.clearDragData();
         }
       });
