@@ -1,30 +1,24 @@
-import {NgModule} from '@angular/core';
-import {EffectsModule} from '@ngrx/effects';
-import {AuthEffects} from './auth.effects';
-import {QuoteEffects} from './quote.effects';
-import {ProjectEffects} from './project.effects';
-import {TaskListEffects} from './task-list.effects';
-import {TaskEffects} from './task.effects';
-import {UserEffects} from './user.effects';
-
-export const effects = {
-  auth: AuthEffects,
-  quote: QuoteEffects,
-  projects: ProjectEffects,
-  tasklists: TaskListEffects,
-  tasks: TaskEffects,
-  users: UserEffects
-};
+import { NgModule } from '@angular/core';
+import { EffectsModule } from '@ngrx/effects';
+import { AuthEffects } from './auth.effects';
+import { QuoteEffects } from './quote.effects';
+import { ProjectEffects } from './project.effects';
+import { TaskListEffects } from './task-list.effects';
+import { TaskEffects } from './task.effects';
+import { TaskHistoryEffects } from './task-history.effects';
+import { UserEffects } from './user.effects';
 
 @NgModule({
   imports: [
-    EffectsModule.run(effects.auth),
-    EffectsModule.run(effects.quote),
-    EffectsModule.run(effects.projects),
-    EffectsModule.run(effects.tasklists),
-    EffectsModule.run(effects.tasks),
-    EffectsModule.run(effects.users),
+    EffectsModule.forRoot([
+      AuthEffects,
+      QuoteEffects,
+      ProjectEffects,
+      TaskListEffects,
+      TaskEffects,
+      TaskHistoryEffects,
+      UserEffects
+    ])
   ],
 })
-export class AppEffectsModule {
-}
+export class AppEffectsModule { }
