@@ -14,8 +14,7 @@ export class QuoteEffects {
    */
   @Effect()
   quote$: Observable<Action> = this.actions$
-    .ofType(actions.ActionTypes.QUOTE)
-    .map(toPayload)
+    .ofType<actions.QuoteAction>(actions.QUOTE)
     .switchMap(() => this.quoteService
       .getQuote()
       .map(quote => new actions.QuoteSuccessAction(quote))
