@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {ProjectListComponent} from './project-list';
+import {ProjectListComponent} from './containers/project-list';
 
 const routes: Routes = [
   {path: '', component: ProjectListComponent}
