@@ -1,12 +1,13 @@
-import {NgModule} from '@angular/core';
-import {AuthService} from './auth.service';
-import {ProjectService} from './project.service';
-import {QuoteService} from './quote.service';
-import {TaskListService} from './task-list.service';
-import {TaskService} from './task.service';
-import {AuthGuardService} from './auth-guard.service';
-import {UserService} from './user.service';
-import {MyCalService} from './my-cal.service';
+import { NgModule } from '@angular/core';
+import { AuthService } from './auth.service';
+import { ProjectService } from './project.service';
+import { QuoteService } from './quote.service';
+import { TaskListService } from './task-list.service';
+import { TaskService } from './task.service';
+import { AuthGuardService } from './auth-guard.service';
+import { UserService } from './user.service';
+import { MyCalService } from './my-cal.service';
+import { TaskHistoryService } from './task-history.service'
 
 export {
   AuthGuardService,
@@ -15,6 +16,7 @@ export {
   QuoteService,
   TaskListService,
   TaskService,
+  TaskHistoryService,
   UserService,
   MyCalService,
 }
@@ -31,6 +33,7 @@ export class ServicesModule {
         QuoteService,
         TaskListService,
         TaskService,
+        TaskHistoryService,
         UserService,
         MyCalService,
       ]
