@@ -10,7 +10,7 @@ export class ProjectService {
   private headers = new HttpHeaders()
     .set('Content-Type', 'application/json');
 
-  constructor(@Inject('BASE_CONFIG') private config,
+  constructor(@Inject('BASE_CONFIG') private config: {uri: string},
               private http: HttpClient) {
     // this.headers.append('X-LC-Id', config.LCId);
     // this.headers.append('X-LC-Key', config.LCKey);
