@@ -6,7 +6,7 @@ import {Quote} from '../domain';
 @Injectable()
 export class QuoteService {
   // private uri: string = 'https://api.hzy.pw/saying/v1/ciba';
-  constructor(@Inject('BASE_CONFIG') private config,
+  constructor(@Inject('BASE_CONFIG') private config: {uri: string},
               private http: HttpClient) {
   }
 
