@@ -9,7 +9,7 @@ export class TaskService {
   private headers = new HttpHeaders()
     .set('Content-Type', 'application/json');
 
-  constructor(@Inject('BASE_CONFIG') private config,
+  constructor(@Inject('BASE_CONFIG') private config: {uri: string},
               private http: HttpClient) {
   }
 
@@ -66,15 +66,15 @@ export class TaskService {
 
   getByLists(lists: TaskList[]): Observable<Task[]> {
     return Observable.from(lists)
-      .mergeMap(list => this.get(list.id))
-      .reduce((tasks, t) => [...tasks, ...t], []);
+      .mergeMap((list: TaskList) => this.get(<string>list.id))
+      .reduce((tasks: Task[], t: Task[]) => [...tasks, ...t], []);
   }
 
-  moveAll(srcListId, targetListId): Observable<Task[]> {
+  moveAll(srcListId: string, targetListId: string): Observable<Task[]> {
     return this.get(srcListId)
-      .mergeMap(tasks => Observable.from(tasks))
-      .mergeMap(task => this.move(task.id, targetListId))
-      .reduce((arrTasks, t) => {
+      .mergeMap((tasks: Task[]) => Observable.from(tasks))
+      .mergeMap((task: Task) => this.move(<string>task.id, targetListId))
+      .reduce((arrTasks: Task[], t: Task) => {
         return [...arrTasks, t];
       }, []);
   }
