@@ -2,111 +2,81 @@ import {NgModule} from '@angular/core';
 import {ReactiveFormsModule, FormsModule} from '@angular/forms';
 import {CommonModule} from '@angular/common';
 import {
-  MdToolbarModule,
-  MdSidenavModule,
-  MdAutocompleteModule,
-  MdButtonModule,
-  MdButtonToggleModule,
-  MdCardModule,
-  MdCheckboxModule,
-  MdChipsModule,
-  MdDatepickerModule,
-  MdDialogModule,
-  MdGridListModule,
-  MdIconModule,
-  MdInputModule,
-  MdListModule,
-  MdMenuModule,
-  MdNativeDateModule,
-  MdProgressBarModule,
-  MdRadioModule,
-  MdSelectModule,
-  MdTabsModule,
-  MdTooltipModule,
-  MdSelectionModule,
-  MdSlideToggleModule,
+  MatToolbarModule,
+  MatSidenavModule,
+  MatAutocompleteModule,
+  MatButtonModule,
+  MatButtonToggleModule,
+  MatCardModule,
+  MatCheckboxModule,
+  MatChipsModule,
+  MatDatepickerModule,
+  MatDialogModule,
+  MatGridListModule,
+  MatIconModule,
+  MatInputModule,
+  MatListModule,
+  MatMenuModule,
+  MatNativeDateModule,
+  MatProgressBarModule,
+  MatRadioModule,
+  MatSelectModule,
+  MatTabsModule,
+  MatTooltipModule,
+  MatSlideToggleModule,
 } from '@angular/material';
-import {ConfirmDialogComponent} from './confirm-dialog';
-import {ImageListSelectComponent} from './image-list-select';
-import {ChipsListComponent} from './chips-list';
-import {IdentityInputComponent} from './identity-input';
-import {AreaListComponent} from './area-list';
-import {AgeInputComponent} from './age-input';
+import {FlexLayoutModule} from '@angular/flex-layout';
+import {ConfirmDialogComponent} from './components/confirm-dialog';
+import {ImageListSelectComponent} from './components/image-list-select';
+import {ChipsListComponent} from './components/chips-list';
+import {IdentityInputComponent} from './components/identity-input';
+import {AreaListComponent} from './components/area-list';
+import {AgeInputComponent} from './components/age-input';
 import {DirectivesModule} from '../directives';
 
+const MATERIAL_MODULES = [
+  MatToolbarModule, MatSidenavModule, MatAutocompleteModule, MatButtonModule,
+  MatButtonToggleModule, MatCardModule, MatCheckboxModule, MatChipsModule,
+  MatDatepickerModule, MatDialogModule, MatGridListModule, MatIconModule,
+  MatInputModule, MatListModule, MatMenuModule, MatNativeDateModule,
+  MatProgressBarModule, MatRadioModule, MatSelectModule, MatTabsModule,
+  MatTooltipModule, MatSlideToggleModule,
+];
+
+const MODULES = [
+  ...MATERIAL_MODULES,
+  CommonModule,
+  ReactiveFormsModule,
+  FormsModule,
+  FlexLayoutModule,
+  DirectivesModule
+];
+
+const DECLARATIONS = [
+  ConfirmDialogComponent,
+  ImageListSelectComponent,
+  ChipsListComponent,
+  IdentityInputComponent,
+  AreaListComponent,
+  AgeInputComponent,
+];
+
+const EXPORT_COMPONENTS = [
+  ConfirmDialogComponent,
+  ImageListSelectComponent,
+  ChipsListComponent,
+  IdentityInputComponent,
+  AreaListComponent,
+  AgeInputComponent,
+];
+
 @NgModule({
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    FormsModule,
-    MdInputModule,
-    MdSelectModule,
-    MdButtonToggleModule,
-    MdCardModule,
-    MdButtonModule,
-    MdDialogModule,
-    MdGridListModule,
-    MdListModule,
-    MdMenuModule,
-    MdIconModule,
-    MdProgressBarModule,
-    MdCheckboxModule,
-    MdChipsModule,
-    MdDatepickerModule,
-    MdNativeDateModule,
-    MdRadioModule,
-    MdAutocompleteModule,
-    MdTabsModule,
-    MdTooltipModule,
-    MdToolbarModule,
-    MdSidenavModule,
-    MdSlideToggleModule,
-    MdSelectionModule,
-    DirectivesModule,
-  ],
+  imports: MODULES,
   exports: [
-    CommonModule,
-    ReactiveFormsModule,
-    FormsModule,
-    MdInputModule,
-    MdSelectModule,
-    MdButtonToggleModule,
-    MdCardModule,
-    MdButtonModule,
-    MdDialogModule,
-    MdGridListModule,
-    MdListModule,
-    MdMenuModule,
-    MdIconModule,
-    MdProgressBarModule,
-    MdCheckboxModule,
-    MdChipsModule,
-    MdDatepickerModule,
-    MdNativeDateModule,
-    MdRadioModule,
-    MdAutocompleteModule,
-    MdTabsModule,
-    MdTooltipModule,
-    MdSelectionModule,
-    MdToolbarModule,
-    MdSidenavModule,
-    MdSlideToggleModule,
-    DirectivesModule,
-    ConfirmDialogComponent,
-    ImageListSelectComponent,
-    ChipsListComponent,
-    IdentityInputComponent,
-    AreaListComponent,
-    AgeInputComponent,
+    ...MODULES,
+    ...EXPORT_COMPONENTS
   ],
-  declarations: [
-    ConfirmDialogComponent,
-    ImageListSelectComponent,
-    ChipsListComponent,
-    IdentityInputComponent,
-    AreaListComponent,
-    AgeInputComponent,
-  ],
+  declarations: DECLARATIONS,
   entryComponents: [
     ConfirmDialogComponent
   ]
