@@ -1,7 +1,7 @@
-import {TaskVM} from './task.vm';
+import { TaskVM } from './task.vm';
 
 export interface TaskListVM {
-  id: string | null;
+  id?: string | null;
   name: string;
   projectId: string;
   order: number;
