@@ -3,7 +3,12 @@ import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 
 import {AppModule} from './app/app.module';
 import {environment} from './environments/environment';
-import { hmrBootstrap } from './hmr';
+import {hmrBootstrap} from './hmr';
+import './rxjs.imports';
+/**
+ * HammerJS needed to import here to make universal build works
+ */
+import 'hammerjs';
 
 if (environment.production) {
   enableProdMode();
